backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
factory-boy==3.3.0
Faker==24.7.1
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
